refactor: remove ext/console/01_colors.js (#18927)
