+++ ext/console/01_console.js
@@ -119,7 +119,16 @@ const {
   SafeMapIterator,
   ArrayBufferPrototype,
 } = primordials;
-import * as colors_ from "ext:deno_console/01_colors.js";
+
+let noColor = false;
+
+function setNoColor(value) {
+  noColor = value;
+}
+
+function getNoColor() {
+  return noColor;
+}
 
 // Don't use 'blue' not visible on cmd.exe
 const styles = {
@@ -3031,7 +3040,7 @@ function inspectArgs(args, inspectOptions = {}) {
   if (ctx.maxArrayLength === null) ctx.maxArrayLength = Infinity;
   if (ctx.maxStringLength === null) ctx.maxStringLength = Infinity;
 
-  const noColor = colors_.getNoColor();
+  const noColor = getNoColor();
   const first = args[0];
   let a = 0;
   let string = "";
@@ -3146,7 +3155,7 @@ const timerMap = new SafeMap();
 const isConsoleInstance = Symbol("isConsoleInstance");
 
 function getConsoleInspectOptions() {
-  const color = !colors_.getNoColor();
+  const color = !getNoColor();
   return {
     ...getDefaultInspectOptions(),
     colors: color,
@@ -3597,9 +3606,11 @@ export {
   formatNumber,
   formatValue,
   getDefaultInspectOptions,
+  getNoColor,
   inspect,
   inspectArgs,
   quoteString,
+  setNoColor,
   styles,
   wrapConsole,
 };
