+++ ext/console/internal.d.ts
@@ -3,7 +3,7 @@
 /// <reference no-default-lib="true" />
 /// <reference lib="esnext" />
 
-declare module "ext:deno_console/02_console.js" {
+declare module "ext:deno_console/01_console.js" {
   function createFilteredInspectProxy<TObject>(params: {
     object: TObject;
     keys: (keyof TObject)[];
